Remove pre-go1.8 code
