+++ .travis.yml
@@ -1,10 +1,10 @@
 language: go
 
 go:
-  - 1.6.x
-  - 1.7.x
   - 1.8.x
   - 1.9.x
+  - 1.10.x
+  - 1.11.x
 
 go_import_path: google.golang.org/appengine
